extract DisplayLink to separated file
